Fix predicate traversal bug in getSubjectFromPredObject
We iterate starting from the object, but the method did not pursue the
edges in the "reverse" order (because the triple is written "forward"
subject predicate(s) object). This commit makes this go the other way,
and removes a cache of traversed edges. Not sure removing the cache was
the right thing, but its working...